Script update
getting more comfortable working within factory functions and modules. moved some code and changed a few things to make more sense. code now stores x's and o's within the array in the appropriate spot for the future win condition code to check if there is a winning combination or not. will work on that next